+++ src/service/good.ts
@@ -1,4 +1,4 @@
-import { createGood, updateGood } from "api/goods"
+import { createGood, updateGood } from "api/order/goods"
 import { queryClient } from "api/queryClient"
 import { useMutation } from "react-query"
