+++ src/type/order/good.ts
@@ -5,3 +5,8 @@ export type Good = {
   name: string
   description?: string
 }
+
+export type GoodInOrder = Good & {
+  quantity: number
+  amount: number
+}
