--- src/type/order/goodInOrder.ts
@@ -1,6 +1,0 @@
-import { Good } from "type/order/good"
-
-export type GoodInOrder = Good & {
-  quantity: number
-  amount: number
-}
